+++ utilities/transactions/lock/range/range_tree/locktree/wfg.cc
@@ -129,7 +129,7 @@ bool wfg::cycle_exists_from_txnid(TXNID txnid,
 // graph, or the function f returns non-zero.
 void wfg::apply_nodes(int (*fn)(TXNID id, void *extra), void *extra) {
   int r = 0;
-  size_t n_nodes = m_nodes.size();
+  uint32_t n_nodes = m_nodes.size();
   for (size_t i = 0; i < n_nodes && r == 0; i++) {
     node *n;
     r = m_nodes.fetch(i, &n);
